pyproject.toml
[project]
name = "backend-pueblo-madera"
version = "0.1.0"
description = ""
authors = [
    {name = "PardoDev",email = "carlos.daniel.pardo.b@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "coreapi (>=2.3.3,<3.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
